completed backend apis, mailers and jobs

+++ controllers/product.js
@@ -1,9 +1,20 @@
-const productModel = require("../models/product");
-const cloudinary = require("../config/cloudinary");
-const fs = require("fs");
+const productModel = require('../models/product');
+const auctionModel = require('../models/auction');
+const bidModel = require('../models/bid');
+const purchaseModel = require('../models/purchase');
+const userModel = require('../models/user');
+const cloudinary = require('../config/cloudinary');
+const fs = require('fs');
+const { response } = require('express');
 
 const getProducts = async (req, res) => {
-  const products = await productModel.find();
+  if (req.userData.role !== 'admin') {
+    return res.status(401).send({
+      message: 'You do not have permission to perform this action.',
+    });
+  }
+  const products = await productModel.find().populate('currentAuction');
+  console.log(req);
 
   return res.status(200).json({
     success: true,
@@ -14,13 +25,16 @@ const getProducts = async (req, res) => {
 
 const getProduct = async (req, res) => {
   const { id } = req.params;
-  const product = await productModel.findOne({ _id: id });
+  const product = await productModel
+    .findOne({ _id: id })
+    .populate('currentAuction');
+  const seller = await userModel.findOne({ _id: product.seller });
 
   try {
     if (product) {
       return res.status(200).json({
         success: true,
-        data: product,
+        data: { product, seller },
       });
     }
   } catch (error) {
@@ -32,32 +46,39 @@ const getProduct = async (req, res) => {
 };
 
 const createProduct = async (req, res) => {
-  const { seller, name, description, minimumBidAmount, status } = req.body;
-  const uploader = async (path) => await cloudinary.uploads(path, "Images");
+  if (req.userData.role === 'buyer') {
+    return res.status(401).send({
+      message: 'You do not have permission to perform this action.',
+    });
+  }
+  const { seller, name, description, minimumBidAmount } = req.body;
+  const uploader = async (path) => await cloudinary.uploads(path, 'Images');
   try {
-    if (req.method === "POST") {
+    if (req.method === 'POST') {
       const urls = [];
       const files = req.files;
-      for (const file of files) {
-        const { path } = file;
-        const newPath = await uploader(path);
-        urls.push(newPath);
-        fs.unlinkSync(path);
+      if (files) {
+        for (const file of files) {
+          const { path } = file;
+          const newPath = await uploader(path);
+          urls.push(newPath);
+          fs.unlinkSync(path);
+        }
       }
       const product = await new productModel({
         seller: seller,
         name: name,
         description: description,
         minimumBidAmount: minimumBidAmount,
-        status: status,
+        status: 'available',
         images: urls,
       });
-      product.save();
-
-      return res.status(201).json({
-        success: true,
-        message: "product created successfully",
-        data: product,
+      product.save().then((response) => {
+        return res.status(201).json({
+          success: true,
+          message: 'product created successfully',
+          data: response,
+        });
       });
     } else {
       return res.status(405).json({
@@ -74,20 +95,28 @@ const createProduct = async (req, res) => {
 
 const updateProduct = async (req, res) => {
   const { seller, name, description, minimumBidAmount, status } = req.body;
-  const { id } = req.params.id;
+  const { id } = req.params;
   let product = await productModel.findById(id);
-  const uploader = async (path) => await cloudinary.uploads(path, "Images");
+  if (
+    product.seller.toHexString() !== req.userData.id &&
+    req.userData.role !== 'admin'
+  ) {
+    response.status(401).send({ message: 'Unauthorized User' });
+  }
+  const uploader = async (path) => await cloudinary.uploads(path, 'Images');
 
   try {
     if (product) {
-      if (req.method === "PUT") {
+      if (req.method === 'PUT') {
         const urls = [];
         const files = req.files;
-        for (const file of files) {
-          const { path } = file;
-          const newPath = await uploader(path);
-          urls.push(newPath);
-          fs.unlinkSync(path);
+        if (files) {
+          for (const file of files) {
+            const { path } = file;
+            const newPath = await uploader(path);
+            urls.push(newPath);
+            fs.unlinkSync(path);
+          }
         }
         product.updateOne(
           {
@@ -106,7 +135,7 @@ const updateProduct = async (req, res) => {
 
         return res.status(201).json({
           success: true,
-          message: "product updated sucessfully",
+          message: 'product updated successfully',
           data: req.body,
         });
       } else {
@@ -117,7 +146,7 @@ const updateProduct = async (req, res) => {
     } else {
       return res.status(400).json({
         success: false,
-        message: "product not found",
+        message: 'product not found',
       });
     }
   } catch (error) {
@@ -129,13 +158,30 @@ const updateProduct = async (req, res) => {
 };
 
 const deleteProduct = async (req, res) => {
-  const { id } = req.params;
-
+  const id = req.params.id;
+  let product = await productModel.findById(id);
+  if (!product) {
+    return res.status(404).json({
+      success: false,
+      message: 'Product not found.',
+    });
+  }
+  if (
+    product.seller.toHexString() !== req.userData.id &&
+    req.userData.role !== 'admin'
+  ) {
+    res.status(401).send({ message: 'Unauthorized User' });
+  }
   try {
+    await bidModel.deleteMany({ product: id });
+    await auctionModel.updateMany(
+      { products: id },
+      { $pull: { products: id } }
+    );
     await productModel.deleteOne({ _id: id });
-    return res.status(410).json({
+    return res.status(200).json({
       success: true,
-      message: "product deleted sucessfully",
+      message: 'product deleted successfully.',
     });
   } catch (error) {
     return res.status(412).send({
@@ -145,10 +191,61 @@ const deleteProduct = async (req, res) => {
   }
 };
 
+const getLiveProducts = async (req, res) => {
+  try {
+    const page = parseInt(req.query.page) || 1;
+    const pageSize = parseInt(req.query.pageSize) || 10;
+    const skip = (page - 1) * pageSize;
+
+    const liveProducts = await productModel.find({ status: 'live' });
+
+    return res.status(200).json({
+      success: true,
+      data: liveProducts,
+      message: 'Live products retrieved successfully.',
+    });
+  } catch (err) {
+    return res.status(412).send({
+      success: false,
+      message: err.message,
+    });
+  }
+};
+
+// const getProductsReport = async (userId) => {
+//   try {
+//     const sellerProducts = await productModel.find({ seller: userId });
+//     return sellerProducts;
+//   } catch (err) {
+//     console.error(err);
+//     throw new Error('Unable to fetch seller products report.');
+//   }
+// };
+
+const getUserProducts = async (userId, role) => {
+  try {
+    if (role === 'seller') {
+      return await productModel
+        .find({ seller: userId })
+        .populate('currentAuction')
+        .populate('seller')
+        .populate('buyer');
+    } else if (role === 'buyer') {
+      return await purchaseModel.find({ buyer: userId });
+    }
+  } catch (err) {
+    console.error(err);
+    throw new Error('Unable to fetch products.');
+  }
+};
+
 module.exports = {
   getProducts,
   getProduct,
   createProduct,
   updateProduct,
   deleteProduct,
+  getLiveProducts,
+  // getProductsReport,
+  getUserProducts,
 };
